refactor: split node renderer batch scheduler
